Start of toolbox and flats....

+++ tsx_cmd_install.sh
@@ -42,27 +42,24 @@ export ROOT_URL='http://127.0.0.1'
 
 if [ "$(uname)" == "Darwin" ]; then
     echo Mac in ${install_dir}
+    if [ "${1}" != "update-only" ]; then
 
-    # install mongodb
-    echo " *******************************"
-    echo "Mongodb - Download and extract"
-    echo " *******************************"
-    curl https://fastdl.mongodb.org/osx/mongodb-osx-ssl-x86_64-4.0.0.tgz -o mongodb-osx-ssl-x86_64-4.0.0.tgz
-    tar -xf mongodb-osx-ssl-x86_64-4.0.0.tgz
-    rm ${install_dir}/mongodb-osx-ssl-x86_64-4.0.0.tgz
+      # install mongodb
+      echo " *******************************"
+      echo "Mongodb - Download and extract"
+      echo " *******************************"
+      curl https://fastdl.mongodb.org/osx/mongodb-osx-ssl-x86_64-4.0.0.tgz -o mongodb-osx-ssl-x86_64-4.0.0.tgz
+      tar -xf mongodb-osx-ssl-x86_64-4.0.0.tgz
+      rm ${install_dir}/mongodb-osx-ssl-x86_64-4.0.0.tgz
 
-    # install nodejs
-    echo " *******************************"
-    echo "nodejs - Download and extract"
-    echo " *******************************"
-    curl https://nodejs.org/dist/v8.11.3/node-v8.11.3-darwin-x64.tar.gz -o node-v8.11.3-darwin-x64.tar.gz
-    tar -xf node-v8.11.3-darwin-x64.tar.gz
-    rm ${install_dir}/node-v8.11.3-darwin-x64.tar.gz
-
-    # install tsx__cmd - assumes already downloaded
-    echo " *******************************"
-    echo " TSX_CMD - Extract" ${1}
-    echo " *******************************"
+      # install nodejs
+      echo " *******************************"
+      echo "nodejs - Download and extract"
+      echo " *******************************"
+      curl https://nodejs.org/dist/v8.11.3/node-v8.11.3-darwin-x64.tar.gz -o node-v8.11.3-darwin-x64.tar.gz
+      tar -xf node-v8.11.3-darwin-x64.tar.gz
+      rm ${install_dir}/node-v8.11.3-darwin-x64.tar.gz
+    fi
     # https://drive.google.com/drive/folders/1yUPU6A0gbBv5UnuSp308lvctY4d6aUtw?usp=sharing
     export PATH=${install_dir}/mongodb-osx-x86_64-4.0.0/bin:${install_dir}/node-v8.11.3-darwin-x64/bin:$PATH
 
@@ -73,24 +70,26 @@ elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
     # install nodejs
     if [ "$(uname -p)" == "aarch64" ]; then
 
-      # install mongodb
-      echo " *******************************"
-      echo "Mongodb - apt-get install"
-      echo " *******************************"
-      # source: https://andyfelong.com/2018/02/update-mongodb-3-6-on-odroid-c2-with-ubuntu-16-04-3-arm64/
-      sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv 2930ADAE8CAF5059EE73BB4B58712A2291FA4AD5
-      echo "deb [ arch=amd64,arm64,ppc64el,s390x ] http://repo.mongodb.com/apt/ubuntu xenial/mongodb-enterprise/3.6 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-enterprise.list
-      sudo apt update
-      sudo apt install g++ build-essential -y
-      sudo apt install mongodb-enterprise -y
+      if [ "${1}" != "update-only" ]; then
+        # install mongodb
+        echo " *******************************"
+        echo "Mongodb - apt-get install"
+        echo " *******************************"
+        # source: https://andyfelong.com/2018/02/update-mongodb-3-6-on-odroid-c2-with-ubuntu-16-04-3-arm64/
+        sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv 2930ADAE8CAF5059EE73BB4B58712A2291FA4AD5
+        echo "deb [ arch=amd64,arm64,ppc64el,s390x ] http://repo.mongodb.com/apt/ubuntu xenial/mongodb-enterprise/3.6 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-enterprise.list
+        sudo apt update
+        sudo apt install g++ build-essential -y
+        sudo apt install mongodb-enterprise -y
 
-      echo " *******************************"
-      echo "nodejs - Download and extract"
-      echo " *******************************"
-      # armv7
-      curl https://nodejs.org/dist/v8.11.3/node-v8.11.3-linux-armv7l.tar.xz -o node-v8.11.3-linux-armv7l.tar.xz
-      tar -xf node-v8.11.3-linux-armv7l.tar.xz
-      rm ${install_dir}/node-v8.11.3-linux-armv7l.tar.xz
+        echo " *******************************"
+        echo "nodejs - Download and extract"
+        echo " *******************************"
+        # armv7
+        curl https://nodejs.org/dist/v8.11.3/node-v8.11.3-linux-armv7l.tar.xz -o node-v8.11.3-linux-armv7l.tar.xz
+        tar -xf node-v8.11.3-linux-armv7l.tar.xz
+        rm ${install_dir}/node-v8.11.3-linux-armv7l.tar.xz
+      fi
       export PATH=${install_dir}/node-v8.11.3-linux-armv7l/bin:$PATH
       # armv8
       #curl https://nodejs.org/dist/v8.11.3/node-v8.11.3-linux-arm64.tar.xz -o node-v8.11.3-linux-arm64.tar.xz
@@ -98,22 +97,24 @@ elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
       #rm ${install_dir}/node-v8.11.3-linux-arm64.tar.xz
     elif [ "$(uname -p)" == "armv7l" ]; then
       # install mongodb
-      echo " *******************************"
-      echo "Mongodb - BUILD - NEED TO BE BUILT FOR ARMv7"
-      echo "run the following (it can take a while on ODroid-XU4)"
-      echo " ./mongod_arm_build.sh"
-      echo " *******************************"
+      if [ "${1}" != "update-only" ]; then
+        echo " *******************************"
+        echo "Mongodb - BUILD - NEED TO BE BUILT FOR ARMv7"
+        echo "run the following (it can take a while on ODroid-XU4)"
+        echo " ./mongod_arm_build.sh"
+        echo " *******************************"
 
-      echo " *******************************"
-      echo "nodejs - Download and extract"
-      echo " *******************************"
-      # armv7
-      curl https://nodejs.org/dist/latest-v6.x/node-v6.14.4-linux-armv7l.tar.gz -o node-v6.14.4-linux-armv7l.tar.gz
-      tar -xf node-v6.14.4-linux-armv7l.tar.gz
-      rm ${install_dir}/node-v6.14.4-linux-armv7l.tar.gz
-      # FIX for odroid
-      rm -rf ~/.node-gyp
-      npm install -g node-gyp
+        echo " *******************************"
+        echo "nodejs - Download and extract"
+        echo " *******************************"
+        # armv7
+        curl https://nodejs.org/dist/latest-v6.x/node-v6.14.4-linux-armv7l.tar.gz -o node-v6.14.4-linux-armv7l.tar.gz
+        tar -xf node-v6.14.4-linux-armv7l.tar.gz
+        rm ${install_dir}/node-v6.14.4-linux-armv7l.tar.gz
+        # FIX for odroid - 2018-11-16 no longer needed
+        #rm -rf ~/.node-gyp
+        #npm install -g node-gyp
+      fi
       export PATH=${install_dir}/node-v6.14.4-linux-armv7l/bin:$PATH
     else
       echo NO NODEJS supported
@@ -125,21 +126,28 @@ else
     exit 5
 fi
 
-# install tsx__cmd - assumes already downloaded
-echo " *******************************"
-echo " TSX_CMD - Extract" ${1}
-echo " *******************************"
-# https://drive.google.com/drive/folders/1yUPU6A0gbBv5UnuSp308lvctY4d6aUtw?usp=sharing
+cd ${install_dir}
 if [ "${1}" != "update-only" ]; then
+  # install tsx__cmd - assumes already downloaded
+  echo " *******************************"
+  echo " TSX_CMD - Extract" ${1}
+  echo " *******************************"
+  # https://drive.google.com/drive/folders/1yUPU6A0gbBv5UnuSp308lvctY4d6aUtw?usp=sharing
     tar -xf ${1}
 fi
 
+cd ${install_dir}/bundle/programs/server
+if [ "${1}" != "update-only" ]; then
+  echo " *******************************"
+  echo " TSX_CMD - update npm installs"
+  echo " *******************************"
+  npm install amdefine ansi-styles chalk escape-string-regexp has-ansi promise source-map strip-ansi type-of ansi-regex asap eachline meteor-promise semver source-map-support supports-color underscore
+fi
 echo " *******************************"
 echo " TSX_CMD - fix for fibers deploy"
 echo " *******************************"
-cd ${install_dir}/bundle/programs/server
 npm uninstall fibers
-npm install fibers@2.0.2 amdefine ansi-styles chalk escape-string-regexp has-ansi promise source-map strip-ansi type-of ansi-regex asap eachline meteor-promise semver source-map-support supports-color underscore
+npm install fibers
 cd ${install_dir}
 
 echo " *******************************"
